Sécurisation des endpoints admin + ajout de l'insertion multiple d'articles via /bulk

+++ backend-user/src/main/java/com/ecommerce/ajc/controller/ArticleClientController.java
@@ -23,7 +23,7 @@ public class ArticleClientController {
     }
 
     @GetMapping("/{id}")
-    public Article getArticleById(@PathVariable Long id) {
+    public Article getArticleById(@PathVariable Integer  id) {
         return articleService.getArticleById(id);
     }
